+++ crates/objc2/CHANGELOG.md
@@ -94,6 +94,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `ffi::objc_ivar` is merged into `runtime::Ivar`.
   - `ffi::BOOL` and constants is merged into `runtime::Bool`.
 * Deprecated `ffi::id`. Use `AnyObject` instead.
+* Deprecated `NSObjectProtocol::is_kind_of`, use `isKindOfClass` or the new
+  `AnyObject::downcast` method instead.
 
 ### Removed
 * **BREAKING**: Removed the `ffi::SEL` and `ffi::objc_selector` types. Use